Убрал лишние импорты, перенёс логин и логаут в views/auth
