+++ source/api/urls/auth.py
@@ -1,15 +1,10 @@
-from django.conf import settings
 from django.urls import path
-from django.contrib.auth import views as auth_views
-from django.conf.urls.static import static
 
 from source.api.views.auth import (
     RegisterView,
     LoginView,
     LogoutView
 )
-
-from source.api.views.pages import change_password_page
 
 urlpatterns = [
     path('register/', RegisterView.as_view(), name='register'),
